viewBox optimization, change "svg-14" transform

+++ Ellipse.h
@@ -5,14 +5,14 @@
 
 class my_ellipse : public SVGElement {
 private:
-  int cx, cy, rx, ry;
+  double cx, cy, rx, ry;
   double stroke_width, stroke_opacity, fill_opacity;
   Color stroke, fill;
   string transform;
 
 public:
-  my_ellipse(string name, string transform, int cx = 0, int cy = 0, int rx = 1,
-             int ry = 1, Color stroke = Color(0, 0, 0), double stroke_width = 1,
+  my_ellipse(string name, string transform, double cx = 0, double cy = 0, double rx = 1,
+      double ry = 1, Color stroke = Color(0, 0, 0), double stroke_width = 1,
              double stroke_opacity = 1, Color fill = Color(0, 0, 0),
              double fill_opacity = 0)
       : cx(cx), cy(cy), rx(rx), ry(ry), stroke(stroke), fill(fill),
@@ -27,10 +27,11 @@ public:
     Pen pen(stroke_color, this->stroke_width);
     SolidBrush brush(fill_color);
 
-    graphic.FillEllipse(&brush, this->cx - this->rx, this->cy - this->ry,
-                        this->rx * 2, this->ry * 2);
-    graphic.DrawEllipse(&pen, this->cx - this->rx, this->cy - this->ry,
-                        this->rx * 2, this->ry * 2);
+	graphic.FillEllipse(&brush, static_cast<REAL>(this->cx - this->rx), static_cast<REAL>(this->cy - this->ry),
+		static_cast<REAL>(this->rx * 2), static_cast<REAL>(this->ry * 2));
+	graphic.DrawEllipse(&pen, static_cast<REAL>(this->cx - this->rx), static_cast<REAL>(this->cy - this->ry),
+		static_cast<REAL>(this->rx * 2), static_cast<REAL>(this->ry * 2));
+
   }
 
   ~my_ellipse() override {}
